:recycle: Refactor: 일정 삭제 API 이동
