+++ application/external-api-v2/src/main/java/com/namo/spring/application/external/api/schedule/controller/ScheduleController.java
@@ -52,17 +52,4 @@ public class ScheduleController {
         return ResponseDto.onSuccess("알림 수정 성공");
     }
 
-    @Operation(summary = "일정 삭제", description = "일정을 삭제합니다.")
-    @ApiErrorCodes(value = {
-            NOT_FOUND_SCHEDULE_FAILURE,
-            NOT_SCHEDULE_PARTICIPANT,
-            NOT_SCHEDULE_OWNER
-    })
-    @DeleteMapping("/{scheduleId}")
-    public ResponseDto<String> deleteSchedule(@Parameter(description = "일정 ID") @PathVariable Long scheduleId,
-                                              @AuthenticationPrincipal SecurityUserDetails member) {
-        scheduleUsecase.deleteSchedule(scheduleId, member);
-        return ResponseDto.onSuccess("일정 삭제 성공");
-    }
-
 }
