+++ application/external-api-v2/src/main/java/com/namo/spring/application/external/api/schedule/usecase/ScheduleUsecase.java
@@ -1,16 +1,6 @@
 package com.namo.spring.application.external.api.schedule.usecase;
 
-import com.namo.spring.application.external.api.schedule.service.ScheduleManageService;
-import com.namo.spring.core.common.code.status.ErrorStatus;
-import com.namo.spring.db.mysql.domains.schedule.entity.Schedule;
-import com.namo.spring.db.mysql.domains.schedule.exception.ScheduleException;
-import com.namo.spring.db.mysql.domains.schedule.service.ScheduleService;
-import com.namo.spring.db.mysql.domains.schedule.type.ParticipantRole;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
 import com.namo.spring.application.external.api.notification.service.NotificationManageService;
-import com.namo.spring.application.external.api.schedule.converter.ScheduleConverter;
 import com.namo.spring.application.external.api.schedule.dto.ScheduleRequest;
 import com.namo.spring.application.external.api.schedule.dto.ScheduleResponse;
 import com.namo.spring.application.external.api.schedule.service.ParticipantManageService;
@@ -18,8 +8,9 @@ import com.namo.spring.application.external.api.user.service.MemberManageService
 import com.namo.spring.application.external.global.common.security.authentication.SecurityUserDetails;
 import com.namo.spring.db.mysql.domains.schedule.entity.Participant;
 import com.namo.spring.db.mysql.domains.user.entity.Member;
-
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import static com.namo.spring.application.external.api.schedule.converter.ScheduleResponseConverter.toScheduleSummaryDto;
 
@@ -29,7 +20,6 @@ public class ScheduleUsecase {
     private final MemberManageService memberManageService;
     private final NotificationManageService notificationManageService;
     private final ParticipantManageService participantManageService;
-    private final ScheduleManageService scheduleManageService;
 
     @Transactional(readOnly = true)
     public ScheduleResponse.ScheduleSummaryDto getScheduleSummary(Long memberId, Long scheduleId) {
@@ -43,11 +33,5 @@ public class ScheduleUsecase {
         Member member = memberManageService.getActiveMember(memberInfo.getUserId());
         notificationManageService.updateOrCreateScheduleReminderNotification(scheduleId, member,
                 dto.getReminderTrigger());
-    }
-
-    @Transactional
-    public void deleteSchedule(Long scheduleId, SecurityUserDetails memberInfo) {
-        Schedule schedule= scheduleManageService.getSchedule(scheduleId);
-        scheduleManageService.deleteSchedule(schedule, memberInfo.getUserId());
     }
 }
